+++ com.unity.postprocessing/PostProcessing/Runtime/Effects/AmbientOcclusion.cs
@@ -196,6 +196,7 @@ namespace UnityEngine.Rendering.PostProcessing
 
                 state &= SystemInfo.supportsComputeShaders
                     && !RuntimeUtilities.isAndroidOpenGL
+                    && !RuntimeUtilities.isWebGL
 #if UNITY_2023_2_OR_NEWER
                     && SystemInfo.IsFormatSupported(GraphicsFormat.R32_SFloat, GraphicsFormatUsage.Render)
                     && SystemInfo.IsFormatSupported(GraphicsFormat.R16_SFloat, GraphicsFormatUsage.Render)